legacy/service-department/requirements.txt
boto3==1.34.72
cerberus==1.3.5
dynaconf==3.2.1
py-moneyed==3.0
structlog==23.1.0
SQLAlchemy==2.0.26
sqlalchemy-utils==0.41.1
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
python-ulid==2.2.0
requests==2.31.0
ory-client==1.2.11
Jinja2==3.1.2
twilio==9.2.3